Secure maintenance routes

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\App;
 use App\Http\Controllers\TracemapController;
 
 // Route principale qui redirige vers la page d'accueil des tracemaps
@@ -19,21 +20,25 @@ Route::post('/tracemaps/ajax', [TracemapController::class, 'storeAjax'])->name('
 
 
 
-//route pour faire artisan migrate
-Route::get('/migrate', function () {
-    Artisan::call('optimize:clear');
-    //script pour supprimer les fichier du dossier storage/app/public/tracemaps
-    File::deleteDirectory(storage_path('app/public/tracemaps'));
+if (App::environment('local')) {
+    Route::middleware('auth')->group(function () {
+        // Route pour lancer les migrations
+        Route::get('/migrate', function () {
+            Artisan::call('optimize:clear');
+            // Supprime les fichiers du dossier storage/app/public/tracemaps
+            File::deleteDirectory(storage_path('app/public/tracemaps'));
 
-    Artisan::call('storage:link');
-    Artisan::call('migrate:fresh --seed');
+            Artisan::call('storage:link');
+            Artisan::call('migrate:fresh --seed');
 
-    return 'Migration terminée';
-});
+            return 'Migration terminée';
+        });
 
-Route::get('/queue', function () {
-    Artisan::call('queue:work');
+        Route::get('/queue', function () {
+            Artisan::call('queue:work');
 
-    return 'queue';
-});
+            return 'queue';
+        });
+    });
+}
 Route::get('logs', [\Rap2hpoutre\LaravelLogViewer\LogViewerController::class, 'index']);
